requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
<<<<<<< HEAD
brotli==1.2.0
=======
>>>>>>> e896eb3e5376e3e7b6caed4b35ff0bc0df2f0ec7
cachetools==5.5.2
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.44.1
colorama==0.4.6
cs50==9.4.0
<<<<<<< HEAD
cssselect2==0.8.0
Deprecated==1.2.18
distro==1.9.0
et_xmlfile==2.0.0
=======
Deprecated==1.2.18
>>>>>>> e896eb3e5376e3e7b6caed4b35ff0bc0df2f0ec7
Faker==37.4.0
filelock==3.16.1
Flask==3.1.0
flask-cors==6.0.1
Flask-Limiter==3.12
flask-ngrok==0.0.25
fonttools==4.60.1
frozenlist==1.6.0
fsspec==2024.12.0
geographiclib==2.0
geopy==2.4.1
google==3.0.0
<<<<<<< HEAD
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.189.0
google-auth==2.40.3
google-auth-httplib2==0.3.0
google-cloud-core==2.4.3
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-generativeai==0.8.6
=======
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.3.0
google-crc32c==1.7.1
>>>>>>> e896eb3e5376e3e7b6caed4b35ff0bc0df2f0ec7
google-resumable-media==2.7.2
google_search_results==2.4.2
googleapis-common-protos==1.70.0
greenlet==3.2.0
groq==1.0.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.31.2
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
limits==5.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
<<<<<<< HEAD
metatrader5==5.0.5509
MouseInfo==0.1.3
=======
>>>>>>> e896eb3e5376e3e7b6caed4b35ff0bc0df2f0ec7
multidict==6.4.3
newsapi-python==0.2.7
ngrok==1.4.0
numpy==2.2.1
<<<<<<< HEAD
openpyxl==3.1.5
=======
>>>>>>> e896eb3e5376e3e7b6caed4b35ff0bc0df2f0ec7
ordered-set==4.1.0
packaging==24.2
pandas==2.3.1
pillow==11.2.1
playwright==1.58.0
propcache==0.3.1
proto-plus==1.26.1
<<<<<<< HEAD
protobuf==5.29.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.11.0
pyee==13.0.0
PyGetWindow==0.0.9
=======
protobuf==6.32.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
>>>>>>> e896eb3e5376e3e7b6caed4b35ff0bc0df2f0ec7
Pygments==2.19.2
PyJWT==2.10.1
PyMsgBox==2.0.1
pyparsing==3.3.2
pyperclip==1.11.0
pyphen==0.17.2
PyRect==0.2.0
PyScreeze==1.0.1
pyTelegramBotAPI==4.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.0
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
PyYAML==6.0.2
qrcode==8.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9.1
safetensors==0.4.5
scipy==1.17.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
tabulate==0.9.0
termcolor==3.0.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.1
twilio==9.6.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
waitress==3.0.2
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
<<<<<<< HEAD
wikipedia==1.4.0
=======
>>>>>>> e896eb3e5376e3e7b6caed4b35ff0bc0df2f0ec7
wrapt==1.17.3
xlsxwriter==3.2.5
yarl==1.20.0
zopfli==0.4.0
gunicorn